Update startup_test_env.py
fix when db_sync doesn't exist initially also

+++ startup_test_env.py
@@ -33,7 +33,8 @@ else:
             else:
                 shutil.rmtree(entry.path)
 
-shutil.rmtree('./run_time/db_sync')
+if os.path.exists('./run_time/db_sync'):
+    shutil.rmtree('./run_time/db_sync')
 time.sleep(10)
 print('Copying src files to mount point')
 shutil.copytree('./src/db_sync','./run_time/db_sync')
